refactor/full-frontend: +minor fixes
