+++ gambitto-frontend/src/widgets/ChessCard/ChessCardWidget.tsx
@@ -1,4 +1,4 @@
-import styles from "./ChessCard.module.scss";
+import styles from "./ChessCardWidget.module.scss";
 import React from "react";
 import Button from "../../components/Button/Button";
 import Card from "../../components/Card/Card";
